Implement database nuking

+++ payup-tm/app_client/common/services/testHelper.service.js
@@ -7,26 +7,32 @@
     // nuke: delete all users from database
     var nuke = function() {
       // Make a delete request.
-      return $http.delete('/api/nukeDB', {
-        headers: {
-          Authorization: 'Bearer ' + authentication.getToken()
-        }
-      }).then(
-        function success(response) {
-          console.log(response.data);
+      if (authentication.isLoggedIn()) {
+        return $http.delete('/api/nukeDB', {
+          headers: {
+            Authorization: 'Bearer ' + authentication.getToken()
+          }
+        }).then(
+          function success(response) {
+            console.log(response.data);
+          });
+      } else {
+        return new Promise(function(resolve) {
+          resolve(false);          
         });
+      }
     };
     
     // fill: fill database with testing data
     var fill = function(idUser, contactData) {
       // Make a post request with empty body.
-        return $http.post('/api/fillDB', {
-            headers: {
-                Authorization: 'Bearer ' + authentication.getToken()
-            }
-        }).then(function success(response) {
-           console.log(response.data); 
-        });
+      return $http.post('/api/fillDB', {
+          headers: {
+              Authorization: 'Bearer ' + authentication.getToken()
+          }
+      }).then(function success(response) {
+         console.log(response.data); 
+      });
     };
     
     // Return object containing the functions offered by this service.
